+++ service/src/main/java/org/ehrbase/service/ValidationServiceImp.java
@@ -41,17 +41,18 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.apache.commons.collections4.CollectionUtils;
 import org.ehrbase.api.exception.InternalServerException;
+import org.ehrbase.api.exception.ObjectNotFoundException;
 import org.ehrbase.api.exception.UnprocessableEntityException;
 import org.ehrbase.api.exception.ValidationException;
+import org.ehrbase.api.service.TemplateService;
 import org.ehrbase.api.service.ValidationService;
 import org.ehrbase.openehr.sdk.response.dto.ContributionCreateDto;
-import org.ehrbase.openehr.sdk.terminology.openehr.TerminologyService;
 import org.ehrbase.openehr.sdk.util.rmconstants.RmConstants;
 import org.ehrbase.openehr.sdk.validation.ConstraintViolation;
 import org.ehrbase.openehr.sdk.validation.ConstraintViolationException;
 import org.ehrbase.openehr.sdk.validation.LocatableValidator;
 import org.ehrbase.openehr.sdk.validation.terminology.ExternalTerminologyValidation;
-import org.ehrbase.openehr.sdk.validation.terminology.ItemStructureVisitor;
+import org.ehrbase.openehr.sdk.validation.terminology.TerminologyValidationVisitor;
 import org.ehrbase.openehr.sdk.validation.webtemplate.FastRMObjectValidator;
 import org.ehrbase.openehr.sdk.webtemplate.model.WebTemplate;
 import org.ehrbase.service.validation.ValidationProperties;
@@ -72,9 +73,8 @@ public class ValidationServiceImp implements ValidationService {
 
     private static final Pattern NAMESPACE_PATTERN = Pattern.compile("[a-zA-Z][a-zA-Z0-9-_:/&+?]*");
 
-    private final KnowledgeCacheServiceImp knowledgeCacheService;
+    private final TemplateService templateService;
 
-    private final TerminologyService terminologyService;
     private final boolean folderValidationEnabled;
 
     private final ThreadLocal<LocatableValidator> locatableValidator;
@@ -82,13 +82,11 @@ public class ValidationServiceImp implements ValidationService {
     private final Map<String, RMPathQuery> rmPathQueryCache = new ConcurrentHashMap<>();
 
     public ValidationServiceImp(
-            KnowledgeCacheServiceImp knowledgeCacheService,
-            TerminologyService terminologyService,
+            TemplateService templateService,
             ValidationProperties validationProperties,
             ObjectProvider<ExternalTerminologyValidation> objectProvider,
             @Value("${cache.validation.useSharedRMPathQueryCache:true}") boolean sharedAqlQueryCache) {
-        this.knowledgeCacheService = knowledgeCacheService;
-        this.terminologyService = terminologyService;
+        this.templateService = templateService;
         this.folderValidationEnabled = validationProperties.validateFolders();
 
         boolean disableStrictValidation = !validationProperties.validateRmConstraints();
@@ -167,9 +165,9 @@ public class ValidationServiceImp implements ValidationService {
     private void check(String templateID, Composition composition) {
         WebTemplate webTemplate;
         try {
-            webTemplate = knowledgeCacheService.getInternalTemplate(templateID);
-        } catch (IllegalArgumentException e) {
-            throw new UnprocessableEntityException(e.getMessage());
+            webTemplate = templateService.getInternalTemplate(templateID);
+        } catch (ObjectNotFoundException e) {
+            throw new UnprocessableEntityException(e.getMessage(), e);
         }
 
         // Validate the composition based on WebTemplate
@@ -179,12 +177,7 @@ public class ValidationServiceImp implements ValidationService {
         }
 
         // check code phrases against terminologies
-        try {
-            ItemStructureVisitor itemStructureVisitor = new ItemStructureVisitor(terminologyService);
-            itemStructureVisitor.validate(composition);
-        } catch (ReflectiveOperationException e) {
-            throw new InternalServerException(e);
-        }
+        new TerminologyValidationVisitor().validate(composition);
     }
 
     private static void compositionMandatoryProperty(Object value, String attribute) {