+++ service/src/test/java/org/ehrbase/service/ValidationServiceTest.java
@@ -21,9 +21,9 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.ehrbase.test.fixtures.EhrStatusFixture.ehrStatus;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.when;
 
 import com.nedap.archie.rm.RMObject;
 import com.nedap.archie.rm.archetyped.Archetyped;
@@ -53,7 +53,6 @@ import com.nedap.archie.rm.support.identification.TerminologyId;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -63,6 +62,7 @@ import java.util.function.Consumer;
 import org.apache.commons.io.IOUtils;
 import org.apache.xmlbeans.XmlException;
 import org.ehrbase.api.exception.ValidationException;
+import org.ehrbase.api.service.TemplateService;
 import org.ehrbase.api.service.ValidationService;
 import org.ehrbase.api.util.ContributionUtils;
 import org.ehrbase.openehr.sdk.response.dto.ContributionCreateDto;
@@ -71,7 +71,6 @@ import org.ehrbase.openehr.sdk.test_data.composition.CompositionTestDataCanonica
 import org.ehrbase.openehr.sdk.test_data.contribution.ContributionTestDataCanonicalJson;
 import org.ehrbase.openehr.sdk.test_data.ehr.EhrTestDataCanonicalJson;
 import org.ehrbase.openehr.sdk.test_data.operationaltemplate.OperationalTemplateTestData;
-import org.ehrbase.openehr.sdk.util.functional.Try;
 import org.ehrbase.openehr.sdk.validation.ConstraintViolation;
 import org.ehrbase.openehr.sdk.validation.ConstraintViolationException;
 import org.ehrbase.openehr.sdk.validation.terminology.ExternalTerminologyValidation;
@@ -85,41 +84,36 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.Mockito;
 import org.openehr.schemas.v1.OPERATIONALTEMPLATE;
-import org.openehr.schemas.v1.TemplateDocument;
 import org.springframework.beans.factory.ObjectProvider;
 
 class ValidationServiceTest {
 
-    private final KnowledgeCacheServiceImp knowledgeCacheService = mock();
+    private final TemplateService templateService = mock();
 
     private final ValidationProperties serverConfig = new ValidationProperties(true, true, true);
 
     private final ObjectProvider<ExternalTerminologyValidation> objectProvider = mock();
 
-    private class NopTerminologyValidation implements ExternalTerminologyValidation {
+    private static class NopTerminologyValidation implements ExternalTerminologyValidation {
 
         private final ConstraintViolation err = new ConstraintViolation("Terminology validation is disabled");
 
-        public Try<Boolean, ConstraintViolationException> validate(TerminologyParam param) {
-            return Try.failure(new ConstraintViolationException(List.of(err)));
+        public ConstraintViolation validate(TerminologyParam param) {
+            return err;
         }
 
         public boolean supports(TerminologyParam param) {
             return false;
         }
-
-        public List<DvCodedText> expand(TerminologyParam param) {
-            return Collections.emptyList();
-        }
     }
 
-    private final ValidationService spyService = spy(new ValidationServiceImp(
-            knowledgeCacheService, new TerminologyServiceImp(), serverConfig, objectProvider, false));
+    private final ValidationService spyService =
+            spy(new ValidationServiceImp(templateService, serverConfig, objectProvider, false));
 
     @BeforeEach
     void setUp() {
-        Mockito.reset(knowledgeCacheService, objectProvider, spyService);
-        doReturn(new NopTerminologyValidation()).when(objectProvider).getIfAvailable();
+        Mockito.reset(templateService, objectProvider, spyService);
+        when(objectProvider.getIfAvailable()).thenReturn(new NopTerminologyValidation());
     }
 
     private ValidationService service() {
@@ -258,7 +252,7 @@ class ValidationServiceTest {
         OperationalTemplateTestData templateData = OperationalTemplateTestData.findByTemplateId(templateID);
         WebTemplate webTemplate = loadWebTemplate(templateData);
 
-        doReturn(webTemplate).when(knowledgeCacheService).getInternalTemplate(templateID);
+        when(templateService.getInternalTemplate(templateID)).thenReturn(webTemplate);
         service().check(composition);
     }
 
@@ -292,7 +286,7 @@ class ValidationServiceTest {
         OperationalTemplateTestData templateData = OperationalTemplateTestData.findByTemplateId(templateID);
         WebTemplate webTemplate = loadWebTemplate(templateData);
 
-        doReturn(webTemplate).when(knowledgeCacheService).getInternalTemplate(templateID);
+        when(templateService.getInternalTemplate(templateID)).thenReturn(webTemplate);
         ValidationService service = service();
 
         assertThatThrownBy(() -> service.check(composition)).isInstanceOf(ConstraintViolationException.class);
@@ -610,13 +604,12 @@ class ValidationServiceTest {
 
     private static WebTemplate loadWebTemplate(OperationalTemplateTestData data) {
         return webTemplates.computeIfAbsent(data, d -> {
-            TemplateDocument document;
-            try (var in = d.getStream()) {
-                document = TemplateDocument.Factory.parse(in);
-            } catch (IOException | XmlException e) {
+            OPERATIONALTEMPLATE template;
+            try {
+                template = TemplateService.buildOperationalTemplate(d.getStream());
+            } catch (XmlException e) {
                 throw new RuntimeException(e);
             }
-            OPERATIONALTEMPLATE template = document.getTemplate();
             return new OPTParser(template).parse();
         });
     }